PrivyLex/requirements.txt
# requirements must be compatible with python 3.13.3
pyfiglet==1.0.2
borsh-construct==0.1.0
Flask==2.3.2
python-dotenv==1.0.0
openai==1.3.0
httpx==0.24.1
PyPDF2==3.0.1
python-docx==1.1.2
Flask-CORS==4.0.0 